LayoutTests/imported/w3c/web-platform-tests/scroll-animations/scroll-timelines/finish-animation-expected.txt


Harness Error (TIMEOUT), message = null

PASS Finishing an animation with a zero playback rate throws
FAIL Finishing an animation seeks to the end time assert_equals: 'actual' unit type must be 'percent' for "After finishing, the currentTime should be set to the end of the active duration" expected (string) "percent" but got (undefined) undefined
FAIL Finishing an animation with a current time past the effect end jumps back to the end promise_test: Unhandled rejection with value: object "TypeError: Type error"
TIMEOUT Finishing a reversed animation jumps to zero time Test timed out
NOTRUN Finishing a reversed animation with a current time less than zero makes it jump back to zero
NOTRUN Finishing an animation while play-pending resolves the pending task immediately
NOTRUN Finishing an animation attached to inactive timeline while play-pending doesn't resolves the pending task
NOTRUN Finishing an animation resolves the finished promise synchronously
NOTRUN A pending ready promise is resolved and not replaced when the animation is finished
NOTRUN Finishing an animation fires finish event on orphaned element
NOTRUN Finishing a canceled animation sets the current and start times
NOTRUN Finishing idle animation produces correct state and fires finish event.
NOTRUN Finishing idle animation attached to inactive timeline pauses the animation.
NOTRUN Finishing running animation produces correct state and fires finish event.
NOTRUN Finishing running animation attached to inactive timeline pauses the animation.
NOTRUN Finishing a paused animation resolves the start time
NOTRUN Finishing a pause-pending animation resolves the pending task immediately and update the start time
NOTRUN Finishing a pause-pending animation with negative playback rate resolves the pending task immediately
NOTRUN Finishing an animation during an aborted pause makes it finished immediately
NOTRUN A pending playback rate should be applied immediately when an animation is finished
NOTRUN An exception should be thrown if the effective playback rate is zero
